+++ pydantic_factories/__init__.py
@@ -1,5 +1,10 @@
 # flake8: noqa
 from .exceptions import ConfigurationError
+from .extensions import (
+    BeanieDocumentFactory,
+    BeaniePersistenceHandler,
+    OdmanticModelFactory,
+)
 from .factory import ModelFactory
-from .fields import SubFactory, Use
+from .fields import Ignore, Require, Use
 from .protocols import AsyncPersistenceProtocol, SyncPersistenceProtocol