+++ pydantic_factories/extensions/__init__.py
@@ -1,2 +1,3 @@
 # flake8: noqa
 from .beanie_odm import BeanieDocumentFactory, BeaniePersistenceHandler
+from .odmantic_odm import OdmanticModelFactory