+++ pydantic_factories/extensions/odmantic_odm.py
@@ -2,8 +2,8 @@ from typing import TypeVar
 
 from pydantic import BaseModel
 
-from pydantic_factories import ModelFactory
-from pydantic_factories.fields import Ignored
+from pydantic_factories.factory import ModelFactory
+from pydantic_factories.fields import Ignore
 
 try:
     from odmantic import EmbeddedModel, Model
@@ -16,4 +16,4 @@ T = TypeVar("T", Model, EmbeddedModel)
 
 
 class OdmanticModelFactory(ModelFactory[T]):
-    id = Ignored()
+    id = Ignore()