+++ CHANGELOG.md
@@ -20,4 +20,4 @@
 [0.3.4b]
 - add support for forward refs
 - add support for odmantic
-- add Ignored and BuildKwarg fields
+- add Ignore and Require fields